Move ShortcutMenu into own folder, start styling
